Movies component refactory || Add/Edit component enchancements
